requirements.txt
gunicorn
efficientnet==1.1.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
h5py==3.11.0
Jinja2==3.1.3
keras==3.3.2
Keras-Applications==1.0.8
Markdown==3.6
MarkupSafe==2.1.5
numpy==1.26.4
pillow==10.3.0
protobuf==4.25.3
python-dateutil==2.9.0.post0
requests==2.31.0
gdown==5.1.0
Werkzeug==3.0.2
tensorflow==2.16.1